fix(architecture): move remote check to proper layer
Move remote check logic from commands layer to orchestra layer to fix
architecture violation. Label utilities moved to clients layer to allow
reuse across KERNEL and COMMAND_ADAPTER categories.

Architecture changes:
- Create clients/label_utils.py for shared label utilities
- Create orchestra/remote_check.py for orchestra remote check
- Update factory to use orchestra remote check (respects layer boundaries)
- Keep commands/check_support.py as thin wrapper
- Add backward compatibility re-exports

Fixes circular dependency: commands → domain → orchestra → commands
Fixes upward dependency: orchestra (KERNEL) → commands (COMMAND_ADAPTER)

All 37 modularity tests pass.

Part of Issue #2381.
