+++ src/vibe3/clients/__init__.py
@@ -21,6 +21,13 @@ if TYPE_CHECKING:
     )
     from vibe3.clients.github_issues_ops import parse_blocked_by, parse_linked_issues
     from vibe3.clients.github_labels import GhIssueLabelPort, IssueLabelPort
+    from vibe3.clients.label_utils import (
+        LabelAnomaly,
+        collect_label_anomalies,
+        has_manager_assignee,
+        normalize_assignees,
+        normalize_labels,
+    )
     from vibe3.clients.merged_pr_cache import MergedPRCache
     from vibe3.clients.protocols.backend import BackendProtocol
     from vibe3.clients.protocols.flow import FlowReader, FlowStatePort
@@ -66,17 +73,22 @@ _LAZY_IMPORTS = {
     "GitHubClientProtocol": "vibe3.clients.protocols.github",
     "GitPathProtocol": "vibe3.clients.protocols.git",
     "IssueLabelPort": "vibe3.clients.github_labels",
+    "LabelAnomaly": "vibe3.clients.label_utils",
     "MergedPRCache": "vibe3.clients.merged_pr_cache",
     "RecentPRCache": "vibe3.clients.recent_pr_cache",
     "SerenaClient": "vibe3.clients.serena_client",
     "SQLiteClient": "vibe3.clients.sqlite_client",
     "TriggerableRoleDefinitionProtocol": "vibe3.clients.protocols.role",
     "check_runtime_asset": "vibe3.clients.runtime_assets",
+    "collect_label_anomalies": "vibe3.clients.label_utils",
     "count_references": "vibe3.clients.serena_client",
     "extract_function_names": "vibe3.clients.serena_client",
     "find_repo_root": "vibe3.clients.git_client",
     "get_store": "vibe3.clients.store_context",
+    "has_manager_assignee": "vibe3.clients.label_utils",
     "init_schema": "vibe3.clients.sqlite_schema",
+    "normalize_assignees": "vibe3.clients.label_utils",
+    "normalize_labels": "vibe3.clients.label_utils",
     "parse_blocked_by": "vibe3.clients.github_issues_ops",
     "parse_linked_issues": "vibe3.clients.github_issues_ops",
     "prune_worktrees": "vibe3.clients.git_worktree_ops",
@@ -118,17 +130,22 @@ __all__ = [
     "GitHubClientProtocol",
     "GitPathProtocol",
     "IssueLabelPort",
+    "LabelAnomaly",
     "MergedPRCache",
     "RecentPRCache",
     "SerenaClient",
     "SQLiteClient",
     "TriggerableRoleDefinitionProtocol",
     "check_runtime_asset",
+    "collect_label_anomalies",
     "count_references",
     "extract_function_names",
     "find_repo_root",
     "get_store",
+    "has_manager_assignee",
     "init_schema",
+    "normalize_assignees",
+    "normalize_labels",
     "parse_blocked_by",
     "parse_linked_issues",
     "prune_worktrees",
